requirements.txt
cx_Oracle==8.3.0
cx_Oracle==8.3.0
dbus_python==1.2.18
holidays==0.46
openpyxl==3.1.2
openrouteservice==2.3.3
pandas==2.2.3
Pillow==9.0.1
Pillow==11.0.0
psycopg2==2.9.2
pypdf==5.1.0
pysftp==0.2.9
python_ldap==3.2.0
qfieldcloud_sdk==0.9.0
Requests==2.32.3
setuptools==59.6.0
workalendar==17.0.0
xlsxwriter==3.2.0
